refactor: 문자열 계산기 리팩토링
- 팩토리, 전략패턴 제거
- Expression 내부에서 식을 생성할 수 있도록 변경
- 게산기 내부에서 구분자를 주입받을 수 있도록 설정
