--- src/main/kotlin/calculator/SeparationStrategy.kt
@@ -1,5 +1,0 @@
-package calculator
-
-interface SeparationStrategy {
-    fun separate(input: String): List<String>
-}
